hotfix/http-client-assertion-trait
* trait moved
